Add different module files
